auth: implement code_verifier logic
Also reorganize the AuthorizationCodeGrantFlows code
